src/main/java/ru/javawebinar/topjava/web/meal/MealRestController.java
package ru.javawebinar.topjava.web.meal;

import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
import ru.javawebinar.topjava.model.Meal;
import ru.javawebinar.topjava.to.MealTo;

import java.net.URI;
import java.time.LocalDateTime;
import java.util.List;

@RestController
@RequestMapping(value = MealRestController.URL, produces = MediaType.APPLICATION_JSON_VALUE)
public class MealRestController extends AbstractMealController {
    static final String URL = "/rest/meals";

    @Override
    @GetMapping("/{id}")
    public Meal get(@PathVariable int id) {
        return super.get(id);
    }

    @Override
    @DeleteMapping("/{id}")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void delete(@PathVariable int id) {
        super.delete(id);
    }

    @Override
    @GetMapping
    public List<MealTo> getAll() {
        return super.getAll();
    }

    @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<Meal> createWithLocation(@RequestBody Meal meal) {
        Meal meal1 = super.create(meal);
        URI uriNew = ServletUriComponentsBuilder.fromCurrentContextPath().path(URL + "/{id}").buildAndExpand(meal1.getId()).toUri();
        return ResponseEntity.created(uriNew).body(meal1);
    }

    @Override
    @PutMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_VALUE)
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void update(@RequestBody Meal meal, @PathVariable int id) {
        super.update(meal, id);
    }


    @GetMapping("/filter")
    public List<MealTo> getBetween(@RequestParam(required = false) LocalDateTime startLocalDateTime, @RequestParam(required = false) LocalDateTime endLocalDateTime) {
        return super.getBetween(startLocalDateTime.toLocalDate(), startLocalDateTime.toLocalTime(), endLocalDateTime.toLocalDate(), endLocalDateTime.toLocalTime());
    }
}